Unify all the shared map fixtures into conftest.py
